requirements.txt
biopython>=1.80
pandas>=2.0
numpy>=1.24
fastmcp>=2.0
loguru>=0.7
click>=8.0
tqdm>=4.60
einops
